requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0
streamlit>=1.25.0
Pillow>=10.0.0
numpy>=1.24.0